pyproject.toml
[project]
name = "auto-calendar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.158.0",
    "google-auth>=2.37.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "pydantic>=2.10.5",
    "pydantic-settings>=2.7.1",
    "requests>=2.32.3",
]
